Create KidsCollectingStrawberries
This is something that I was supposed to do with a bunch of FOR loops, but I've failed miserably over and over to realize how to avoid berries being a global variable, so in the end I've given up and decided to just declare a function that calculates the loop inside and only returns a print statement, instead of overwriting the global variable. Now github copilot neatly suggested to globalize my berries inside the function body and for some rookie reason taken his suggestion, which ended up overwriting my berries yet again. Eventually I realized I just need to move berries completely into the function body and that did it. Now it works as intended.